requirements.txt
python-magic-bin==0.4.14
werkzeug==1.0.1
flask==1.1.2
flask_restful==0.3.8
scipy==1.4.1
tensorflow==1.11.0
joblib==0.15.1
keras==2.2.4
MulticoreTSNE==0.1
matplotlib==3.2.1
librosa==0.7.2
scikit-learn==0.20.0
spherecluster==0.1.7
wavefile==1.5
numpy==1.18.4
